#29 Refactor command factories folder
