+++ src/commands/commandList.js
@@ -1,13 +1,13 @@
-const ApproveCommandBuilder = require('../factories/commands/ApproveCommandBuilder')
-const CreateWalletCommandBuilder = require('../factories/commands/CreateWalletCommandBuilder')
-const DepositEthCommandBuilder = require('../factories/commands/DepositEthCommandBuilder')
-const DepositTokenCommandBuilder = require('../factories/commands/DepositTokenCommandBuilder')
-const GetAddressCommandBuilder = require('../factories/commands/GetAddressCommandBuilder')
-const GetAllowanceCommandBuilder = require('../factories/commands/GetAllowanceCommandBuilder')
-const GetBalanceCommandBuilder = require('../factories/commands/GetBalanceCommandBuilder')
-const OrderCancelCommandBuilder = require('../factories/commands/OrderCancelCommandBuilder')
-const OrderSignCommandBuilder = require('../factories/commands/OrderSignCommandBuilder')
-const WithdrawCommandBuilder = require('../factories/commands/WithdrawCommandBuilder')
+const ApproveCommandBuilder = require('../factories/commands/trading-wallet/ApproveCommandBuilder')
+const CreateWalletCommandBuilder = require('../factories/commands/trading-wallet/CreateWalletCommandBuilder')
+const DepositEthCommandBuilder = require('../factories/commands/trading-wallet/DepositEthCommandBuilder')
+const DepositTokenCommandBuilder = require('../factories/commands/trading-wallet/DepositTokenCommandBuilder')
+const GetAddressCommandBuilder = require('../factories/commands/trading-wallet/GetAddressCommandBuilder')
+const GetAllowanceCommandBuilder = require('../factories/commands/trading-wallet/GetAllowanceCommandBuilder')
+const GetBalanceCommandBuilder = require('../factories/commands/trading-wallet/GetBalanceCommandBuilder')
+const OrderCancelCommandBuilder = require('../factories/commands/order/OrderCancelCommandBuilder')
+const OrderSignCommandBuilder = require('../factories/commands/order/OrderSignCommandBuilder')
+const WithdrawCommandBuilder = require('../factories/commands/trading-wallet/WithdrawCommandBuilder')
 
 const approveCommand = ApproveCommandBuilder.build()
 const createWalletCommand = CreateWalletCommandBuilder.build()
