+++ src/factories/commands/order/OrderCancelCommandBuilder.js
@@ -1,10 +1,10 @@
-const logger = require('../../logger')
-const { PrivateKeyService } = require('../../services/PrivateKeyService')
+const logger = require('../../../logger')
+const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
-const OrderCancelCommand = require('../../commands/order/OrderCancelCommand')
-const OrderCancelCommandValidator = require('../../validators/OrderCancelCommandValidator')
-const OrderService = require('../../services/OrderService')
-const PrivateKeyValidator = require('../../validators/PrivateKeyValidator')
+const OrderCancelCommand = require('../../../commands/order/OrderCancelCommand')
+const OrderCancelCommandValidator = require('../../../validators/OrderCancelCommandValidator')
+const OrderService = require('../../../services/OrderService')
+const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
 
 const orderService = new OrderService(logger)
 const privateKeyService = new PrivateKeyService(logger)
