+++ src/factories/commands/trading-wallet/CreateWalletCommandBuilder.js
@@ -1,10 +1,10 @@
-const logger = require('../../logger')
-const { PrivateKeyService } = require('../../services/PrivateKeyService')
+const logger = require('../../../logger')
+const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
-const CreateWalletCommand = require('../../commands/trading-wallet/CreateWalletCommand')
-const CreateWalletCommandValidator = require('../../validators/CreateWalletCommandValidator')
-const GetAllowanceCommandValidator = require('../../validators/GetAllowanceCommandValidator')
-const PrivateKeyValidator = require('../../validators/PrivateKeyValidator')
+const CreateWalletCommand = require('../../../commands/trading-wallet/CreateWalletCommand')
+const CreateWalletCommandValidator = require('../../../validators/CreateWalletCommandValidator')
+const GetAllowanceCommandValidator = require('../../../validators/GetAllowanceCommandValidator')
+const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
 
 const privateKeyService = new PrivateKeyService(logger)
 const privateKeyValidator = new PrivateKeyValidator(logger)
