+++ src/factories/commands/trading-wallet/DepositTokenCommandBuilder.js
@@ -1,10 +1,10 @@
-const logger = require('../../logger')
-const { PrivateKeyService } = require('../../services/PrivateKeyService')
+const logger = require('../../../logger')
+const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
-const DepositTokenCommand = require('../../commands/trading-wallet/DepositTokenCommand')
-const DepositTokenCommandValidator = require('../../validators/DepositTokenCommandValidator')
-const PrivateKeyValidator = require('../../validators/PrivateKeyValidator')
-const TradingWalletServiceBuilder = require('../TradingWalletServiceBuilder')
+const DepositTokenCommand = require('../../../commands/trading-wallet/DepositTokenCommand')
+const DepositTokenCommandValidator = require('../../../validators/DepositTokenCommandValidator')
+const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
+const TradingWalletServiceBuilder = require('../../TradingWalletServiceBuilder')
 
 const privateKeyService = new PrivateKeyService(logger)
 const privateKeyValidator = new PrivateKeyValidator(logger)
