+++ src/factories/commands/trading-wallet/GetAllowanceCommandBuilder.js
@@ -1,9 +1,9 @@
-const logger = require('../../logger')
-const { PrivateKeyService } = require('../../services/PrivateKeyService')
+const logger = require('../../../logger')
+const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
-const GetAllowanceCommand = require('../../commands/token/GetAllowanceCommand')
-const GetAllowanceCommandValidator = require('../../validators/GetAllowanceCommandValidator')
-const PrivateKeyValidator = require('../../validators/PrivateKeyValidator')
+const GetAllowanceCommand = require('../../../commands/token/GetAllowanceCommand')
+const GetAllowanceCommandValidator = require('../../../validators/GetAllowanceCommandValidator')
+const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
 
 const privateKeyService = new PrivateKeyService(logger)
 const privateKeyValidator = new PrivateKeyValidator(logger)
